+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 1.01 [22-01-2026]
+
+**Retail:**  
+- Block chat command usage while in combat  
+- Still register /cd if WA is loaded instead of disabling this Addon  
+
 ## Version 1.00 [21-01-2026]
 
 **Retail:**